+++ packages/atomic/cypress/e2e/result-list/result-components/result-localized-text-selectors.ts
@@ -1,0 +1,9 @@
+import {ResultListSelectors} from '../result-list-selectors';
+
+export const resultLocalizedTextComponent = 'atomic-result-localized-text';
+
+export const ResultLocalizedTextSelectors = {
+  shadow: () => cy.get(resultLocalizedTextComponent),
+  firstInResult: () =>
+    ResultListSelectors.firstResult().find(resultLocalizedTextComponent),
+};