+++ packages/atomic/src/components/common/atomic-result-placeholder/placeholders.tsx
@@ -10,32 +10,16 @@ export interface ResultPlaceholderProps {
   imageSize?: ResultDisplayImageSize;
   display?: ResultDisplayLayout;
   numberOfPlaceholders: number;
-  // TODO: check is still needed
-  isChild?: boolean;
 }
 
-export const ListDisplayResultsPlaceholder: FunctionalComponent<
-  ResultPlaceholderProps
-> = (props) => {
+export const ResultsPlaceholder: FunctionalComponent<ResultPlaceholderProps> = (
+  props
+) => {
   return Array.from({length: props.numberOfPlaceholders}, (_, i) => (
     <atomic-result-placeholder
       key={`placeholder-${i}`}
       density={props.density}
-      display="list"
-      imageSize={props.imageSize!}
-      isChild={props.isChild}
-    ></atomic-result-placeholder>
-  ));
-};
-
-export const GridDisplayResultsPlaceholder: FunctionalComponent<
-  ResultPlaceholderProps
-> = (props) => {
-  return Array.from({length: props.numberOfPlaceholders}, (_, i) => (
-    <atomic-result-placeholder
-      key={`placeholder-${i}`}
-      density={props.density}
-      display="grid"
+      display={props.display || 'list'}
       imageSize={props.imageSize!}
     ></atomic-result-placeholder>
   ));