+++ packages/atomic/src/components/common/result-list/result-list-common.tsx
@@ -1,10 +1,9 @@
-import {Host, h, FunctionalComponent} from '@stencil/core';
+import {h, FunctionalComponent, Fragment} from '@stencil/core';
 import {getFirstFocusableDescendant} from '../../../utils/accessibility-utils';
 import {updateBreakpoints} from '../../../utils/replace-breakpoint';
 import {once} from '../../../utils/utils';
 import {
-  GridDisplayResultsPlaceholder,
-  ListDisplayResultsPlaceholder,
+  ResultsPlaceholder,
   ResultPlaceholderProps,
   TableDisplayResultsPlaceholder,
 } from '../atomic-result-placeholder/placeholders';
@@ -22,6 +21,8 @@ import {
   ResultListRenderer,
 } from './result-list-common-interface';
 import {TableDisplayResults} from './table-display-results';
+
+export const resultComponentClass = 'result-component';
 
 export class ResultListCommon<SpecificResult extends AnyResult = AnyResult>
   implements ResultListRenderer, ResultListInfo
@@ -73,7 +74,7 @@ export class ResultListCommon<SpecificResult extends AnyResult = AnyResult>
 
   public get listClasses() {
     const classes = getResultDisplayClasses(
-      this.props.getDisplay(),
+      this.props.getLayoutDisplay(),
       this.props.getDensity(),
       this.props.getImageSize()
     );
@@ -111,18 +112,18 @@ export class ResultListCommon<SpecificResult extends AnyResult = AnyResult>
     }
 
     return (
-      <Host>
+      <Fragment>
         {this.props.resultTemplateProvider.hasError && <slot></slot>}
         <div class={`list-wrapper ${this.listClasses}`}>
           <ResultDisplayWrapper
             listClasses={this.listClasses}
-            display={this.props.getDisplay()}
+            display={this.props.getLayoutDisplay()}
           >
             {this.displayPlaceholders && (
-              <ResultsPlaceholder
+              <DisplayResultsPlaceholder
                 numberOfPlaceholders={this.props.getNumberOfPlaceholders()}
                 density={this.props.getDensity()}
-                display={this.props.getDisplay()}
+                display={this.props.getResultDisplay()}
                 imageSize={this.props.getImageSize()}
               />
             )}
@@ -136,7 +137,7 @@ export class ResultListCommon<SpecificResult extends AnyResult = AnyResult>
             )}
           </ResultDisplayWrapper>
         </div>
-      </Host>
+      </Fragment>
     );
   }
 }
@@ -156,16 +157,14 @@ const ResultDisplayWrapper: FunctionalComponent<{
   );
 };
 
-const ResultsPlaceholder: FunctionalComponent<ResultPlaceholderProps> = (
+const DisplayResultsPlaceholder: FunctionalComponent<ResultPlaceholderProps> = (
   props
 ) => {
   switch (props.display) {
     case 'table':
       return <TableDisplayResultsPlaceholder {...props} />;
-    case 'grid':
-      return <GridDisplayResultsPlaceholder {...props} />;
     default:
-      return <ListDisplayResultsPlaceholder {...props} />;
+      return <ResultsPlaceholder {...props} />;
   }
 };
 
@@ -176,7 +175,7 @@ const ResultListDisplay: FunctionalComponent<ResultListDisplayProps> = (
     return null;
   }
 
-  switch (props.getDisplay()) {
+  switch (props.getLayoutDisplay()) {
     case 'table':
       return <TableDisplayResults {...props} />;
     case 'grid':