+++ packages/atomic/src/components/insight/atomic-insight-interface/atomic-insight-interface.tsx
@@ -24,6 +24,7 @@ import {
   buildInsightEngine,
   buildInsightResultsPerPage,
 } from '..';
+import {loadFieldActions} from '@coveo/headless/insight';
 
 const FirstInsightRequestExecutedFlag = 'firstInsightRequestExecuted';
 export type InsightInitializationOptions = InsightEngineConfiguration;
@@ -73,7 +74,7 @@ export class AtomicInsightInterface
   /**
    * The language assets path. By default, this will be a relative URL pointing to `./lang`.
    *
-   * @example /mypublicpath/languages
+   * Example: "/mypublicpath/languages"
    *
    */
   @Prop({reflect: true}) public languageAssetsPath = './lang';
@@ -81,7 +82,7 @@ export class AtomicInsightInterface
   /**
    * The icon assets path. By default, this will be a relative URL pointing to `./assets`.
    *
-   * @example /mypublicpath/icons
+   * Example: "/mypublicpath/icons"
    *
    */
   @Prop({reflect: true}) public iconAssetsPath = './assets';
@@ -108,7 +109,6 @@ export class AtomicInsightInterface
 
   public connectedCallback() {
     this.store.setLoadingFlag(FirstInsightRequestExecutedFlag);
-    this.updateFieldsToInclude();
   }
 
   private initResultsPerPage() {
@@ -120,11 +120,12 @@ export class AtomicInsightInterface
     });
   }
 
-  private updateFieldsToInclude() {
+  public registerFieldsToInclude() {
     if (this.fieldsToInclude) {
-      this.store.set(
-        'fieldsToInclude',
-        this.fieldsToInclude.split(',').map((field) => field.trim())
+      this.engine!.dispatch(
+        loadFieldActions(this.engine!).registerFieldsToInclude(
+          this.fieldsToInclude.split(',').map((field) => field.trim())
+        )
       );
     }
   }
@@ -182,10 +183,6 @@ export class AtomicInsightInterface
     this.commonInterfaceHelper.onAnalyticsChange();
   }
 
-  render() {
-    return this.engine && <slot></slot>;
-  }
-
   public get bindings(): InsightBindings {
     return {
       engine: this.engine!,
@@ -215,8 +212,13 @@ export class AtomicInsightInterface
 
   private async internalInitialization(initEngine: () => void) {
     await this.commonInterfaceHelper.onInitialization(initEngine);
+    this.registerFieldsToInclude();
     this.store.unsetLoadingFlag(FirstInsightRequestExecutedFlag);
     this.initResultsPerPage();
     this.initialized = true;
+  }
+
+  render() {
+    return this.engine && <slot></slot>;
   }
 }