+++ packages/atomic/src/components/insight/atomic-insight-result-list/atomic-insight-result-list.tsx
@@ -61,7 +61,7 @@ export class AtomicInsightResultList
   @State() private resultTemplateRegistered = false;
   @State() public error!: Error;
 
-  @FocusTarget<InsightBindings>() nextNewResultTarget!: FocusTargetController;
+  @FocusTarget() nextNewResultTarget!: FocusTargetController;
 
   /**
    * The spacing of various elements in the result list, including the gap between results, the gap between parts of a result, and the font sizes of different parts in a result.
@@ -116,7 +116,8 @@ export class AtomicInsightResultList
       host: this.host,
       bindings: this.bindings,
       getDensity: () => this.density,
-      getDisplay: () => this.display,
+      getResultDisplay: () => this.display,
+      getLayoutDisplay: () => this.display,
       getImageSize: () => this.imageSize,
       nextNewResultTarget: this.nextNewResultTarget,
       loadingFlag: this.loadingFlag,