+++ packages/atomic/src/components/recommendations/atomic-recs-interface/analytics-config.ts
@@ -1,0 +1,44 @@
+import {
+  AnalyticsPayload,
+  augmentAnalyticsWithAtomicVersion,
+  augmentWithExternalMiddleware,
+  augmentAnalyticsConfigWithDocument,
+} from '../../common/interface/analytics-config';
+import {
+  AnalyticsConfiguration,
+  RecommendationEngineConfiguration,
+} from '@coveo/headless/recommendation';
+
+export function getAnalyticsConfig(
+  recsConfig: RecommendationEngineConfiguration,
+  enabled: boolean
+): AnalyticsConfiguration {
+  const analyticsClientMiddleware = (
+    event: string,
+    payload: AnalyticsPayload
+  ) => augmentAnalytics(event, payload, recsConfig);
+
+  const defaultAnalyticsConfig: AnalyticsConfiguration = {
+    analyticsClientMiddleware,
+    enabled,
+    ...augmentAnalyticsConfigWithDocument(),
+  };
+
+  if (recsConfig.analytics) {
+    return {
+      ...defaultAnalyticsConfig,
+      ...recsConfig.analytics,
+    };
+  }
+  return defaultAnalyticsConfig;
+}
+
+function augmentAnalytics(
+  event: string,
+  payload: AnalyticsPayload,
+  recsConfig: RecommendationEngineConfiguration
+) {
+  let result = augmentWithExternalMiddleware(event, payload, recsConfig);
+  result = augmentAnalyticsWithAtomicVersion(result);
+  return result;
+}