+++ packages/atomic/src/components/recommendations/atomic-recs-interface/atomic-recs-interface.tsx
@@ -1,4 +1,12 @@
-import {RecommendationEngine, LogLevel} from '@coveo/headless/recommendation';
+import {
+  RecommendationEngine,
+  RecommendationEngineConfiguration,
+  LogLevel,
+  loadFieldActions,
+  EcommerceDefaultFieldsToInclude,
+  buildRecommendationEngine,
+  loadRecommendationActions,
+} from '@coveo/headless/recommendation';
 import {
   Component,
   Element,
@@ -7,6 +15,7 @@ import {
   Method,
   Prop,
   Watch,
+  State,
 } from '@stencil/core';
 import i18next, {i18n} from 'i18next';
 import {InitializeEvent} from '../../../utils/initialization-utils';
@@ -15,9 +24,11 @@ import {
   BaseAtomicInterface,
   CommonAtomicInterfaceHelper,
 } from '../../common/interface/interface-common';
+import {getAnalyticsConfig} from './analytics-config';
 import {createAtomicRecsStore, AtomicRecsStore} from './store';
 
 const FirstRecommendationExecutedFlag = 'firstRecommendationExecuted';
+export type RecsInitializationOptions = RecommendationEngineConfiguration;
 export type RecsBindings = CommonBindings<
   RecommendationEngine,
   AtomicRecsStore,
@@ -38,13 +49,35 @@ export class AtomicRecsInterface
 {
   private store = createAtomicRecsStore();
   private commonInterfaceHelper: CommonAtomicInterfaceHelper<RecommendationEngine>;
+  private initialized = false;
 
   @Element() public host!: HTMLAtomicRecsInterfaceElement;
+
+  @State() public error?: Error;
+
+  /**
+   * The recommendation interface [search hub](https://docs.coveo.com/en/1342/).
+   */
+  @Prop({reflect: true}) public searchHub = 'default';
+
+  /**
+   * The [tz database](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones) identifier of the time zone to use to correctly interpret dates in the query expression, facets, and result items.
+   * By default, the timezone will be [guessed](https://day.js.org/docs/en/timezone/guessing-user-timezone).
+   *
+   * Example: "America/Montreal"
+   */
+  @Prop({reflect: true}) public timezone?: string;
 
   /**
    * The recommendation interface headless engine.
    */
   @Prop({mutable: true}) public engine?: RecommendationEngine;
+
+  /**
+   * The recommendation interface [query pipeline](https://docs.coveo.com/en/180/).
+   */
+  @Prop({reflect: true}) public pipeline?: string;
+
   /**
    * Whether analytics should be enabled.
    */
@@ -54,24 +87,26 @@ export class AtomicRecsInterface
    * The recommendation interface i18next instance.
    */
   @Prop() public i18n: i18n = i18next.createInstance();
+
   /**
    * The severity level of the messages to log in the console.
    */
   @Prop({reflect: true}) public logLevel?: LogLevel;
+
   /**
-   * The search interface language.
+   * The recommendation interface language.
    */
   @Prop({reflect: true}) public language = 'en';
 
   /**
-   * TODO: implement fieldsToInclude
+   * A list of non-default fields to include in the query results, separated by commas.
    */
   @Prop({reflect: true}) public fieldsToInclude = '';
 
   /**
    * The language assets path. By default, this will be a relative URL pointing to `./lang`.
    *
-   * @example /mypublicpath/languages
+   * Example: "/mypublicpath/languages"
    *
    */
   @Prop({reflect: true}) public languageAssetsPath = './lang';
@@ -79,7 +114,7 @@ export class AtomicRecsInterface
   /**
    * The icon assets path. By default, this will be a relative URL pointing to `./assets`.
    *
-   * @example /mypublicpath/icons
+   * Example: "/mypublicpath/icons"
    *
    */
   @Prop({reflect: true}) public iconAssetsPath = './assets';
@@ -105,13 +140,42 @@ export class AtomicRecsInterface
   }
 
   /**
+   * Initializes the connection with the headless recommendation engine using options for `accessToken` (required), `organizationId` (required), `renewAccessToken`, and `platformUrl`.
+   */
+  @Method() public initialize(options: RecsInitializationOptions) {
+    return this.internalInitialization(() => this.initEngine(options));
+  }
+
+  /**
    * Initializes the connection with an already preconfigured headless recommendation engine.
-   *
+   * This bypasses the properties set on the component, such as analytics, recommendation, searchHub, language, timezone & logLevel.
    */
   @Method() public initializeWithRecommendationEngine(
     engine: RecommendationEngine
   ) {
     return this.internalInitialization(() => (this.engine = engine));
+  }
+
+  /**
+   *
+   * Fetches new recommendations.
+   */
+  @Method() public async getRecommendations() {
+    if (!this.commonInterfaceHelper.engineIsCreated(this.engine)) {
+      return;
+    }
+
+    if (!this.initialized) {
+      console.error(
+        'You have to wait until the "initialize" promise is fulfilled before executing a search.',
+        this.host
+      );
+      return;
+    }
+
+    this.engine!.dispatch(
+      loadRecommendationActions(this.engine!).getRecommendations()
+    );
   }
 
   @Watch('iconAssetsPath')
@@ -134,9 +198,41 @@ export class AtomicRecsInterface
     this.commonInterfaceHelper.onAnalyticsChange();
   }
 
+  public registerFieldsToInclude() {
+    const fields = EcommerceDefaultFieldsToInclude.concat(
+      this.fieldsToInclude.split(',').map((field) => field.trim())
+    );
+    this.engine!.dispatch(
+      loadFieldActions(this.engine!).registerFieldsToInclude(fields)
+    );
+  }
+
   private async internalInitialization(initEngine: () => void) {
     await this.commonInterfaceHelper.onInitialization(initEngine);
     this.store.unsetLoadingFlag(FirstRecommendationExecutedFlag);
+    this.initialized = true;
+  }
+
+  private initEngine(options: RecsInitializationOptions) {
+    const analyticsConfig = getAnalyticsConfig(options, this.analytics);
+    try {
+      this.engine = buildRecommendationEngine({
+        configuration: {
+          searchHub: this.searchHub,
+          locale: this.language,
+          timezone: this.timezone,
+          pipeline: this.pipeline,
+          ...options,
+          analytics: analyticsConfig,
+        },
+        loggerOptions: {
+          level: this.logLevel,
+        },
+      });
+    } catch (error) {
+      this.error = error as Error;
+      throw error;
+    }
   }
 
   public render() {