+++ packages/atomic/src/components/recommendations/index.ts
@@ -1,0 +1,7 @@
+export {
+  RecommendationEngineConfiguration,
+  Result as RecsResult,
+  ResultTemplateCondition as RecsResultTemplateCondition,
+  ResultTemplate as RecsResultTemplate,
+  ResultTemplatesHelpers as RecsResultTemplatesHelpers,
+} from '@coveo/headless/recommendation';