+++ packages/atomic/src/components/search/atomic-search-interface/atomic-search-interface.tsx
@@ -21,6 +21,7 @@ import {
   loadSearchConfigurationActions,
   loadQueryActions,
   EcommerceDefaultFieldsToInclude,
+  loadFieldActions,
 } from '@coveo/headless';
 import i18next, {i18n} from 'i18next';
 import {InitializeEvent} from '../../../utils/initialization-utils';
@@ -72,6 +73,7 @@ export class AtomicSearchInterface
    * A list of non-default fields to include in the query results, separated by commas.
    */
   @Prop({reflect: true}) public fieldsToInclude = '';
+
   /**
    * The search interface [query pipeline](https://docs.coveo.com/en/180/).
    */
@@ -91,8 +93,7 @@ export class AtomicSearchInterface
    * The [tz database](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones) identifier of the time zone to use to correctly interpret dates in the query expression, facets, and result items.
    * By default, the timezone will be [guessed](https://day.js.org/docs/en/timezone/guessing-user-timezone).
    *
-   * @example
-   * America/Montreal
+   * Example: "America/Montreal"
    */
   @Prop({reflect: true}) public timezone?: string;
 
@@ -129,7 +130,7 @@ export class AtomicSearchInterface
   /**
    * The language assets path. By default, this will be a relative URL pointing to `./lang`.
    *
-   * @example /mypublicpath/languages
+   * Example: "/mypublicpath/languages"
    *
    */
   @Prop({reflect: true}) public languageAssetsPath = './lang';
@@ -137,7 +138,7 @@ export class AtomicSearchInterface
   /**
    * The icon assets path. By default, this will be a relative URL pointing to `./assets`.
    *
-   * @example /mypublicpath/icons
+   * Example: "/mypublicpath/icons"
    *
    */
   @Prop({reflect: true}) public iconAssetsPath = './assets';
@@ -153,7 +154,7 @@ export class AtomicSearchInterface
   public connectedCallback() {
     this.store.setLoadingFlag(FirstSearchExecutedFlag);
     this.updateMobileBreakpoint();
-    this.updateFieldsToInclude();
+    this.initFieldsToInclude();
   }
 
   @Watch('searchHub')
@@ -238,7 +239,7 @@ export class AtomicSearchInterface
 
   /**
    * Initializes the connection with an already preconfigured headless search engine, as opposed to the `initialize` method which will internally create a new search engine instance.
-   *
+   * This bypasses the properties set on the component, such as analytics, searchHub, pipeline, language, timezone & logLevel.
    */
   @Method() public initializeWithSearchEngine(engine: SearchEngine) {
     return this.internalInitialization(() => (this.engine = engine));
@@ -289,25 +290,27 @@ export class AtomicSearchInterface
     };
   }
 
-  public render() {
-    return [
-      this.engine && (
-        <atomic-relevance-inspector
-          bindings={this.bindings}
-        ></atomic-relevance-inspector>
-      ),
-      <slot></slot>,
-    ];
-  }
-
-  private updateFieldsToInclude() {
+  private initFieldsToInclude() {
     const fields = [...EcommerceDefaultFieldsToInclude];
     if (this.fieldsToInclude) {
       fields.push(
         ...this.fieldsToInclude.split(',').map((field) => field.trim())
       );
     }
-    this.store.set('fieldsToInclude', fields);
+    this.store.addFieldsToInclude(fields);
+
+    // TODO: delete v2 when fields-to-include prop on result list removed
+    this.store.onChange('fieldsToInclude', () =>
+      this.registerFieldsToInclude()
+    );
+  }
+
+  public registerFieldsToInclude() {
+    this.engine?.dispatch(
+      loadFieldActions(this.engine!).registerFieldsToInclude(
+        this.store.state.fieldsToInclude
+      )
+    );
   }
 
   private updateMobileBreakpoint() {
@@ -433,8 +436,20 @@ export class AtomicSearchInterface
 
   private async internalInitialization(initEngine: () => void) {
     await this.commonInterfaceHelper.onInitialization(initEngine);
+    this.initFieldsToInclude();
     this.initSearchStatus();
     this.initUrlManager();
     this.initialized = true;
+  }
+
+  public render() {
+    return [
+      this.engine && (
+        <atomic-relevance-inspector
+          bindings={this.bindings}
+        ></atomic-relevance-inspector>
+      ),
+      <slot></slot>,
+    ];
   }
 }